backend/requirements.txt
django>=4.0
djangorestframework
django-cors-headers
python-decouple
pydub
requests
